Add async Get method to ClothingItemApiController
An asynchronous `Get` method was added to the `ClothingItemApiController` in the `ClimFit.Api.Controllers.ApiControllers` namespace. The method is marked with the `[HttpGet]` attribute and overrides a base class implementation. It retrieves a list of clothing items from `_clothingItemService.GetListAsync()`, filters them by the logged-in user's ID using `GetLoggedInUserId()`, and returns the filtered list as an HTTP 200 OK response.